when traversing directories, always skip on files.
With the current impl, we'll do a native call to list files for a given file, which should be a noop, but it's not necessary since we only care about directories.
This change makes it so that consumers don't have to check if a file is a directory by ensuring they all are. It also prevents calls to list when the a file is a file and not a directory

Reviewed at https://rbcommons.com/s/twitter/r/1516/