Fixed compiler output for combined modules
It appeared if specify compiler output as a : separated list of folders. IntelliJ will construct a proper classpath for modules.

Also removed _module from preview because it appeared IJ fails to remove it on the next refresh if there is another module with the same content root.

Testing Done:
Tests are passing.

Reviewed at https://rbcommons.com/s/twitter/r/2214/