Upgrade ./rbt in the intellij-pants-plugin repo (copied from pants).
Upgrade ./rbt in the intellij-pants-plugin repo (verbatim copy from pantsbuild/pants/rbt).

This resolves an apparent issue with older rbtools where ~/.reviewboardrc is honored before ./.reviewboardrc, which manifests itself as the following failure:

    [illuminati intellij-pants-plugin (master)]$ ./rbt patch -c 3134
    Patch is being applied from request 3134 with diff revision  1.
    CRITICAL: Object does not exist (HTTP 404, API Error 100)

This was determined to be due to a ~/.reviewboardrc pointing to an internal RB server which when moved out of the way (or with this patch) works as expected.

Testing Done:
verified expected `./rbt status --debug` behavior before and after the change.

Reviewed at https://rbcommons.com/s/twitter/r/3146/