Disable zooming on project dir on subsequent project resolves
The view focus feature waits for modules to be present before zooming. However since there are existing modules in subsequent project resolves and due to its async nature, there is a chance that the zoom may take place on modules to be replaced, causing intellij to try to find them forever.

This patch disables the zoom on subsequent resolves, plus many cases are from project refresh and zooming again on the initial import dir is not as useful.

Testing Done:
https://travis-ci.org/wisechengyi/intellij-pants-plugin/builds/153639409

Reviewed at https://rbcommons.com/s/twitter/r/4173/