Suggest refresh on project BUILD file changes (#254)
## Problem
BUILD file changes in the project does not notify user to refresh. The issue is more obvious after file changes out of band such as `git pull`.

## Solution
This change adds listener specifically on BUILD files in the project, and notify user to refresh the project if there are any changes. 

For example: 
![screen shot 2017-02-02 at 3 16 25 pm](https://cloud.githubusercontent.com/assets/596358/22572786/99dcafc6-e95a-11e6-8aa0-e5bcf3a909f0.png)