Capture indexing duration on the fly (#265)
## Problem
There was no good way to capture indexing timing on the fly, but only on synthetic benchmark.

## Solution
This change uses `DumbService.DumbModeListener` interface to calculate the duration spent being dumb, i.e. the IDE is indexing.

## Result
In manual test, the results are captured; in unit test, the results are updated by artificially marking and unmarking dumb mode.