refactor pants invocation to only add cli options when appropriate (#318)
The previous logic adds cli options (e.g. `--async`) to the command line if the version of pants supports e.g. `clean-all.async`, regardless of whether the command line contains the `clean-all` task. This makes it harder to construct different command lines with different tasks such as lint, because if `--async` is added without the `clean-all` task, pants throws an error. Instead, we ensure that cli options such as `--async` are only added if they are relevant to one of the tasks.

An alternative would be to make sure the cli options are less broadly scoped (e.g. `--clean-all-async` instead of `--async`), but adding a bunch of options that either go unused or interfere with the desired result sounds like it could cause problems later.