Add assertions for accidental bug fix from #312 (#320)
In #312, a change was made to [`CompileAllTargetsInModuleAction.java`](https://github.com/pantsbuild/intellij-pants-plugin/pull/312/files#diff-41cc874d589fdbd635d9ca9f78149ef9) which stopped a bug that affected usage.

To reproduce the error, take a version of the plugin from any commit before b2a024a52a9287d74944400dc4f27a7a613989ba. Open a pants project A, then open a pants project B in a separate window. Click `Build->Compile Module '...'` with a source file from A open in the editor, which should work correctly. Switch to the window for project B, open a source file from B in the editor, then click `Build->Compile Module '...'`. The name of the module will be correct, but the `./pants compile ...` command will build the module from project A (which can be seen in the pants console -- the targets for project A are selected).

#312 accidentally fixed this (see link to diff at the top of this comment) by not mutating the private `module` field in `CompileAllTargetsInModuleAction`, but I didn't realize this modified any behavior until the bug above came to my attention. This adds multiple assertions to `OSSPantsCompileActionsTest#testCompileAllTargetsInModuleAction` to avoid a regression.