If there are a large number of target specs, pass them to pants via a file to avoid exceeding the OS command line length limit. (#404)
If there are a large number of target specs, pass them to pants via a file to avoid exceeding the OS command line length limit.

Specifically uses a threshold of 100 right now, which is directly in the code as opposed to being defined in a constants file. This seems to be consistent with the rest of the plugin code, which is littered with strings, although there is at least one shared constant `PantsCompileOptionsExecutor.PROJECT_NAME_LIMIT`. In practice, the threshold of 100 should not make a difference to the rest of the plugin code.

Does not currently include any unit tests. In principle it seems like it should be possible to unit test `PantsCompileOptionsExecutor.getPantsExportCommand()`.