Refresh SDK in a background
Previously, the refresh was happening on the EDT thread which
caused errors and freezes because of a process being started inside the
"createPantsJdk" call.

As a result of running in the background, changing the project JDK
(if it wasn't a pants' one) must be scheduled to run in a dispatcher
thread.