Don't drop list of targets while saving settings
Before this commit, the list of the targets were stored in
`PantsProjectSettings::myTargetSpecs` field. Unfortunately, this list was based on the
contents of `myTargetSpecsBox` widget, which was cleared each time
when the "Pants->Settings" window was opened. As a result, the list
of targets was cleared each time during settings save.

The fix to this is just not to clean the list of targets in the
window.

But this led to a new issue. When the list of targets is presented
to the user, it's also possible for him to deselect target. During
save, the `myTargetSpecs` variable was based on the list of selected
targets, so after deselection and save, it was not possible to select
it again.

In order to overcome this, I put two lists to `PantsProjectSettings`
instead of `myTargetSpecs`:

 1. `allAvailableTargetSpecs` - used to build `myTargetSpecsBox` widget
 2. `selectedTargetSpecs` - to inform `pants` which targets it should
operate on

fixes #457