Don't try to reload orphan modules during Pants refresh
Before this change, IntelliJ's ModuleDataService was trying to reload
orphan modules, when the user was trying to refresh pants project and
the refresh caused unload of some modules. The ModuleDataService
thought it is it's job to try to reload modules, because
ExternalModuleType property of the module was unset. Therefore it was
detected by ModuleDataService::computeOrphanData method.

The fix is to make pants set ExternalModuleType property during module
creation, so they're ignored by the ModuleDataService.

fixes #457