Clear projects when doing reimport
The "cleanPantsProjects" method is not called now in
PantsUtil.refreshAllProjects, but in
PantsSystemProjectResolver.resolveProjectInfo,
and it's renamed to "cleanPantsModules".

The problem is that, the "refreshAllProjects" method was not called
each time the reimport occured, it was executed only when the user
clicked Pants->Refresh Pants Projects in menubar

Apart from that, the external project path is now stored as an option,
so we during reimport we removed only the modules that were added with
this path. Therefore, it's possible to import more than one project now.

`setExternalModuleType` in `PantsMetadataService` should be called
because we do not want modules to be catched by ModuleDataService::computeOrphanData