Don't retreive pants options in Event Dispatch Thread
FileChangeTracker methods are always called in Event Dispatch Thread.
Pants Plugin implementation sometimes needs to retrieve pants options
in order to decide, whether to mark project as dirty. Unfortunatly,
Pants Options are retrieved via an external process call, what is
forbidden in the Event Dispatch Thread.

In this commit, the bahvior of the listener is changed in that way,
the pants options are retrieved before the listener is initialized,
on a thread from thread pool, and they're passed to listener's
constructor.

Unfortunately, pants options may be read only when a project
contains any modules. Therefore, it we can't just read them at
`projectOpened` stage, because the project may be initially empty.
That's why we register a ModuleListener in case when the project is
empty.

Fixes #463