Workaround for "Find Usages broken" bug in Scala Plugin
Bug tracker: https://youtrack.jetbrains.com/issue/SCL-16768

Correct initialization of `ScalaCompilerReferenceService` is required for "Find usages" feature to work correcly. Unfortunately, it's broken in that way, it does nothing when there are no modules in the project right after the moment, when the project is opened. This happens when a project is imported, the module creation starts after opening then.