Allow setting project name (#486)
Previously, project name was derived from the targets used for importing it. Now, user can specify it when importing.

The import window looks as follows
![import-project-window-with-name](https://user-images.githubusercontent.com/3709537/73538035-3cf18d80-442a-11ea-8a28-2c538f643f8a.png)

And the resulting structure is:
![imported-modules](https://user-images.githubusercontent.com/3709537/73538181-96f25300-442a-11ea-88b5-d91252b23043.png)

The project view:
![project-list](https://user-images.githubusercontent.com/3709537/73538263-d91b9480-442a-11ea-9799-e41669e6c84b.png)

The name field is being automatically populated if and only if it was not changed by the user.
![autogenerated-name](https://user-images.githubusercontent.com/3709537/73542468-11c06b80-4435-11ea-90c3-6ff45f644b4c.png)

The strategy for generation of project name tries to mimic current approach (with a change of prepending the build root directory name). Since I don't have much experience with using the pants build tool, I could use some suggestions on how those automatically generated names could look like to actually useful.