Save incremental import depth setting
This change was reverted, because we thought it broke CI. But after
reversion the CI is still red with the same errors.

ORIGINAL COMMIT HASH
1f72954011b645a66288544c10818c216e2fb7c8

ORIGINAL MESSAGE

Before this commit, IntelliJ plugin was asking for incremental import
depth each time, even when doing refresh. The depth value was not
saved anywhere.

Now, the incremental import depth parameter is saved alongside other
settings, and may be configured in the same place as others.