Manually close all projects
EditorEvent listeners are unsubscribed from the bus when the project
is closed. Previously it was done by
ExternalSystemImportingTestCase.tearDown. There was also an assertion there
that checked if no listeners leaked. For some reason, since 2020.2,
the subscription list is not cleared correcly during tearDown, so this
commit ads a manual cleanup