Bump to IJ 202.6109.22 (#538)
* Bump to IJ 202.6109.22

Neededd to adapt to new OpenProjectTask
constructor.

external-system-test-api.jar was updated accordingly to
CONTRIBUTING.md guide

* Manually close all projects

EditorEvent listeners are unsubscribed from the bus when the project
is closed. Previously it was done by
ExternalSystemImportingTestCase.tearDown. There was also an assertion there
that checked if no listeners leaked. For some reason, since 2020.2,
the subscription list is not cleared correcly during tearDown, so this
commit ads a manual cleanup

* Increase timeout for cache download

We've been quite close to the limit. It usually took about 500-570
seconds, while the limit was 600

* Trigger CI

* Trigger CI

* Trigger CI