Update for scala plugin 2020.2.717 (#555)
This PR is to test if everything still works with latest scala plugin. Also updates package, as it changed for one of extension points in scala plugin we are using.
We can merge it as is, or wait till the breaking api change is in stable release and revert changes in sh scripts.