Do not call get project.getBasePath
This method may return incorrect value. The case is that the user may
create project literally *anywhere* and tnen import BSP project to
it. `getBasePath` will return path to the directory where the project
was created.

Apart from that, since this commit the `sources-jar` dir is also
recognized as project internal dependency (updated
`isProjectInternalDependency` method)