Merge events happening during python sdk setup (#559)
As per advice from jetbrains, I wrapped calls that mutate project model into mergeRootsChangesDuring. I also removed the unnecessary commit that doubled number of calls.
On project with 1700 modules, the freeze time is down from 2 minutes to below 20 seconds, but freeze is still there.
The update needs to happen in write action and it affects a lot of modules and handlers for roots change will be invoked, so I don't think we can avoid any freezing in current intellij. I added a notification that informs the user for how many modules the update will be invoked and that ide may freeze.